Move tests from test_basic.py to test_pandas.py
